Program 2: Binary Search Tree
In this assignment, you are to write a program that analyzes the text in an input file, counting the number of times each word appears in the text and the line numbers where it appears. 

Specifications & Requirements

For purposes of this assignment, a word is any consecutive string of letters and the apostrophe character, so  don’t  counts as a single word, and  best-selling counts as two words: best and selling.

Your word counts must ignore capitalization, so  the, The, THE, and tHe all increase the count for the word “the” by one.

The input file must not contain numbers.

You must store the words and the number of occurrences and lines on which it occurs in a binary search tree.  Each word occurring in the text can only be stored once in the tree. 

Create a class for the nodes of the tree called WordNode; and call the branches in this structure left and right.

Create a class implementing a binary search tree called  WordTree.  It must also contain the following public methods:

insert:  adds the given word to the tree if it is not already in the tree OR increments the appropriate counter if it is already there, and records the line number,
printTreeInOrder: prints in alphabetical order the words in the input file along with how many times they appear in the file,
countUniqueWords:  returns the number of words currently stored in the tree,
countWords: returns the number of words in the input text,
lastWord: returns the last word in the tree based on alphabetical ordering, 
firstWord: returns the first word in the tree based on alphatetcial ofdering, 
wordLines: returns a vector of line numbers for a word, and
mostFrequent: returns a vector of the most frequently occurring word(s) (i.e., could be more than one)